move files to /lib
