+++ src/index.js
@@ -13,7 +13,7 @@ import { router as nominatimRouter } from './nominatim/index.js';
 import { router as fileRouter } from './file/index.js';
 import { router as geoConfigRouter } from './geoconfig/index.js';
 import { router as realtimeRouter } from './realtime-gtfs/index.js';
-import { loadLookupTables } from './route-linestring/client.js';
+import { loadLookupTables } from './route-linestring/index.js';
 
 if (!satisfies(process.version, '>=18')) {
   console.error('ERROR: bikehopper-web-app requires node v18');
